tools/python/requirements.txt
openpyxl>=3.1.0
PyYAML>=6.0
pandas>=1.5.0
